requirements-server.txt
apscheduler>=3.10,<4.0
psycopg2-binary>=2.9,<3.0
pandas>=2.0,<3.0
pandas-ta>=0.3.14b
requests>=2.31,<3.0
yfinance>=0.2.40
python-dotenv>=1.0,<2.0
anthropic>=0.25,<1.0
google-genai>=1.0,<2.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
